Refactor API structure and add Bootstrap
