+++ app/api/__init__.py
@@ -1,0 +1,3 @@
+from . import auth, files
+
+__all__ = ["auth", "files"]
