+++ app/static/index.html
@@ -3,25 +3,26 @@
 <head>
     <meta charset="utf-8">
     <title>Course Viewer</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
     <style>
-        ul { list-style-type: none; }
+        ul { list-style-type: none; padding-left: 1rem; }
         .dir { font-weight: bold; }
     </style>
 </head>
-<body>
+<body class="container py-4">
 <div id="login">
-    <h2>Login</h2>
-    <input id="user" placeholder="username"><br>
-    <input id="pass" type="password" placeholder="password"><br>
-    <button onclick="login()">Login</button>
-    <div id="login-error"></div>
+    <h2 class="mb-3">Login</h2>
+    <div class="mb-2"><input class="form-control" id="user" placeholder="username"></div>
+    <div class="mb-2"><input class="form-control" id="pass" type="password" placeholder="password"></div>
+    <button class="btn btn-primary" onclick="login()">Login</button>
+    <div class="text-danger mt-2" id="login-error"></div>
 </div>
 <div id="content" style="display:none;">
     <h2>Files</h2>
-    <div id="tree"></div>
-    <iframe id="viewer" style="width:80%; height:500px;"></iframe>
+    <div id="tree" class="mb-3"></div>
+    <iframe id="viewer" class="w-100" style="height:500px;"></iframe>
     <br>
-    <button onclick="nextFile()">Next</button>
+    <button class="btn btn-secondary mt-2" onclick="nextFile()">Next</button>
 </div>
 <script>
 let token = '';
