+++ app/utils/__init__.py
@@ -1,0 +1,17 @@
+from .core import (
+    get_db,
+    scan,
+    ensure_user,
+    build_tree,
+    tokens,
+    FILES_DIR,
+)
+
+__all__ = [
+    "get_db",
+    "scan",
+    "ensure_user",
+    "build_tree",
+    "tokens",
+    "FILES_DIR",
+]
